discord-bot/requirements.txt
discord.py>=2.3.0
anthropic>=0.40.0
openai>=1.50.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
